+++ src/main/java/com/lyf/forum/service/QuestionService.java
@@ -62,7 +62,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public PaginationDTO list(Integer userId, Integer page, Integer size) {
+    public PaginationDTO list(Long userId, Integer page, Integer size) {
         PaginationDTO paginationDTO = new PaginationDTO();
         QuestionExample questionExample=new QuestionExample();
         questionExample.createCriteria().andCreatorEqualTo(userId);
@@ -94,7 +94,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question=questionMapper.selectByPrimaryKey(id);
         if (question==null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -114,9 +114,9 @@ public class QuestionService {
             // 创建
             question.setGmtCreate(System.currentTimeMillis());
             question.setGmtModified(System.currentTimeMillis());
-            question.setCommentCount(0);
-            question.setLikeCount(0);
-            question.setViewCount(0);
+            question.setCommentCount(0l);
+            question.setLikeCount(0l);
+            question.setViewCount(0l);
             questionMapper.insert(question);
         }else {
             // 更新
@@ -130,7 +130,11 @@ public class QuestionService {
         }
     }
 
-    public void incView(Integer id){
+    /**
+     * 增加阅读数
+     * @param id
+     */
+    public void incView(Long id){
         questionExtMapper.incView(id);
     }
 }
